% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombinePackets.R
\name{filesRead}
\alias{filesRead}
\title{filesRead}
\usage{
filesRead(nameFile, form, namesCol)
}
\arguments{
\item{nameFile}{the name of the file, string}

\item{form}{, string '.rds' or '.csv'}
}
\value{
the individual file piece.
}
\description{
A custom function for reading all those files we split.
}
\examples{

x <- filesRead('./namehere.rds', '.rds')

}
