% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePackets.R
\name{makeFiles}
\alias{makeFiles}
\title{makeFiles}
\usage{
makeFiles(file, n, form, j, rows, chosenDir)
}
\arguments{
\item{file, }{a data frame. The data frame that needs converting into many files}

\item{n, }{an integer. The number of mini data frames you'd like to make}

\item{form, }{a string. Can take either '.rds' for R data files or '.csv' for ,csv files.
Will determine the format of the smaller data frames.}

\item{j, }{the number of rows in the big data frame.}

\item{rows, }{the number of rows in each smaller data frame.}

\item{chosenDir, }{a string. Where would you like to put all these small data frames?
Built by makePackets().}
}
\value{
does not return anything
}
\description{
makeFiles
}
\examples{

makeFiles(file, n, form, j, rows, i, chosenDir)

}
