% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePackets.R
\name{makeIndvFiles}
\alias{makeIndvFiles}
\title{makeIndvFiles}
\usage{
makeIndvFiles(file, n, form, j, rows, numberFile, chosenDir)
}
\arguments{
\item{file, }{a data frame. The data frame that needs converting into many files}

\item{n, }{an integer. The number of mini data frames you'd like to make}

\item{form, }{a string. Can take either '.rds' for R data files or '.csv' for ,csv files.
Will determine the format of the smaller data frames.}

\item{j, }{the number of rows in the big data frame.}

\item{rows, }{the number of rows in each smaller data frame.}

\item{numberFile, }{the file number. The smaller data frames are built sequentially so they
can be rebuilt into one data frame in the same order they came in.}

\item{chosenDir, }{a string. Where would you like to put all these small data frames?
Built by makePackets().}
}
\value{
doesn't return anything.
}
\description{
This function does the heavy lifting for the makePackets function.
It builds all the individual files from the larger data frame.
}
\examples{

makeIndvFiles(file, 10, '.csv', 1000, 101, 2, './data')

}
