% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePackets.R
\name{makePackets}
\alias{makePackets}
\title{makePackets}
\usage{
makePackets(file, n, form, chosenDir)
}
\arguments{
\item{file, }{a data frame. The data frame that needs converting into many files}

\item{n, }{an integer. The number of mini data frames you'd like to make}

\item{form, }{a string. Can take either '.rds' for R data files or '.csv' for ,csv files.
Will determine the format of the smaller data frames.}

\item{chosenDir, }{a string. Where would you like to put all these small data frames?
The function will check if that directory exists, if not it will make it.}
}
\value{
does not return.
}
\description{
Turn a very large data frame into an ordered set of smaller ones.
}
\examples{

data <- matrix(runif(1,0,1), nrow = 1000, ncol = 20)
makePackets(data, 10, '.rds', './data')

}
