% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombinePackets.R
\name{recombinePackets}
\alias{recombinePackets}
\title{recombinePackets}
\usage{
recombinePackets(dir, form, namesCol)
}
\arguments{
\item{dir, }{string. Directory where you find the files.}

\item{form, }{string. '.rds' or '.csv'.}

\item{namesCol}{a vector with the names of the columns you're going to call them}
}
\value{
newFile, a data frame with all the bits!
}
\description{
Once you've broken your big data frame into little ones, you need to put
it back together.
}
\examples{

recombinePackts('./', '.rds', c("a", "b", "c"))


}
